Add born cross sections

+++ pydradana/sim_reader.py
@@ -1,8 +1,9 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 # For python 2-3 compatibility
 from __future__ import division, print_function
 
+import numbers
 import os
 import shutil
 
@@ -11,66 +12,77 @@ import uproot
 
 __all__ = ['SimReader']
 
-_cache_dir = '/tmp/uproot_cache'
-try:
-    os.rmdir(os.path.join(_cache_dir, 'order'))
-    raise FileNotFoundError
-except FileNotFoundError:
-    if os.path.exists(_cache_dir):
-        shutil.rmtree(_cache_dir)
-    _disk_cache = uproot.cache.DiskCache.create(50 * 1024**3, _cache_dir)  # 50 GB disk cache
-except OSError as e:
-    if e.errno == 66:
-        _disk_cache = uproot.cache.DiskCache.join(_cache_dir)
+_cache = [None]
+
+
+def setup_cache(path='/tmp', memorysize=0, disksize=0):
+    _cache_dir = os.path.join(path, 'uproot_cache')
+    if disksize > 0:
+        try:
+            os.rmdir(os.path.join(_cache_dir, 'oRDer'))
+            raise FileNotFoundError
+        except FileNotFoundError:
+            if os.path.exists(_cache_dir):
+                shutil.rmtree(_cache_dir)
+            _disk_cache = uproot.cache.DiskCache.create(disksize * 1024**3, _cache_dir)  # 50 GB disk cache
+        except OSError as e:
+            if e.errno == 66:
+                _disk_cache = uproot.cache.DiskCache.join(_cache_dir)
+            else:
+                raise
+        if memorysize > 0:
+            _cache[0] = uproot.cache.MemoryCache(memorysize * 1024**3, spillover=_disk_cache, spill_immediately=True)
+        else:
+            _cache[0] = _disk_cache
     else:
-        raise
-_cache = uproot.cache.MemoryCache(8 * 1024**3, spillover=_disk_cache, spill_immediately=True)
+        if memorysize > 0:
+            _cache[0] = uproot.cache.MemoryCache(memorysize * 1024**3)
+        else:
+            _cache[0] = None
 
 
-class _JaggedArray():
+def get_column(self, index):
+    if not isinstance(index, numbers.Integral):
+        raise TypeError("JaggedArray index must be an integer")
 
-    def __init__(self, jarray):
-        self._jarray = jarray
-        self._data = {}
+    array_at_index = [
+        self.content[self.starts[i] + index] if self.starts[i] + index < self.stops[i] else numpy.nan for i in range(len(self.stops))
+    ]
 
-    def __getitem__(self, key):
-        if not isinstance(key, int):
-            raise KeyError(key)
-        jarray_at_key = [
-            self._jarray.content[self._jarray.starts[i] + key] if self._jarray.starts[i] + key < self._jarray.stops[i] else numpy.nan
-            for i in range(len(self._jarray))
-        ]
-        self._data[key] = numpy.array(jarray_at_key)
-        return self._data[key]
+    return numpy.array(array_at_index)
 
 
-class _Detector():
+uproot.interp.jagged.JaggedArray.get_column = get_column
 
-    def __init__(self, tree, name, var_dict):
+
+class _Detector(object):
+
+    def __init__(self, tree, name, var_dict, start, stop):
         self._tree = tree
         self._name = name
         self._var_dict = var_dict
+        self._start = start
+        self._stop = stop
 
     def __getattr__(self, name):
         if name in self._var_dict:
-            var_data = self._tree.array(self._var_dict[name], cache=_cache)
-            if isinstance(var_data, uproot.interp.jagged.JaggedArray):
-                data = _JaggedArray(var_data)
-            else:
-                data = var_data
+            data = self._tree.array(self._var_dict[name], cache=_cache[0], entrystart=self._start, entrystop=self._stop)
             setattr(self, name, data)
             return data
+
         raise AttributeError(name)
 
 
-class SimReader():
+class SimReader(object):
 
-    def __init__(self, file_name):
+    def __init__(self, file_name, start=0, stop=100000):
+        self.start = start
+        self.stop = stop
         self.open(file_name)
 
     def _add_attrs(self):
         for det, var_dict in self._structure.items():
-            option = _Detector(self._tree, det, var_dict)
+            option = _Detector(self._tree, det, var_dict, self.start, self.stop)
             setattr(self, det, option)
 
     @property
@@ -79,6 +91,7 @@ class SimReader():
 
     def open(self, file_name):
         self._tree = uproot.open(file_name)['T']
+
         # convert the first layer tree structure to a dictionary
         self._structure = {}
         for key in self._tree.keys():
@@ -90,9 +103,31 @@ class SimReader():
             if det not in self._structure:
                 self._structure[det] = {}
             self._structure[det][var] = (det + '.' + var).encode('ascii')
+
         self._add_attrs()
 
+    def get_primary(self):
+        pass
 
-if __name__ == '__main__':
-    r = SimReader('sim.root')
-    print(r.GUN.N)
+    def find_hits_tracking(self, det_name, n_copy):
+        try:
+            det = getattr(self, det_name)
+        except AttributeError:
+            return None
+
+        is_primary = det.PTID.content == 0
+        fired_did_list = numpy.split(det.DID.content, det.DID.stops)[:-1]
+
+        found = []
+        for copyid in range(n_copy):
+            found_in_copy = []
+            for i in range(len(fired_did_list)):
+                if copyid in fired_did_list[i]:
+                    index_list = det.DID.starts[i] + numpy.nonzero(fired_did_list[i] == copyid)[0]  # 0 means x axis
+                    if any(is_primary[index_list]):
+                        found_in_copy.append(index_list[is_primary[index_list]][0])
+                else:
+                    found_in_copy.append(-1)
+            found.append(numpy.array(found_in_copy))
+
+        return found
